# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorDetail'
      success:
        type: boolean
    type: object
  dto.GroupResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.ImportScheduleResponse:
    properties:
      created_subjects:
        items:
          type: string
        type: array
      created_teachers:
        items:
          type: string
        type: array
      errors:
        items:
          type: string
        type: array
      imported_count:
        type: integer
    type: object
  dto.LessonResponse:
    properties:
      ends_at:
        type: string
      id:
        type: string
      queue_id:
        type: string
      room:
        $ref: '#/definitions/dto.RoomResponse'
      starts_at:
        type: string
      subject:
        $ref: '#/definitions/dto.SubjectResponse'
      teacher:
        $ref: '#/definitions/dto.TeacherResponse'
      type:
        type: string
    type: object
...
  dto.QueueMinimalInfo:
    properties:
      id:
        type: string
      status:
        type: string
      subject:
        $ref: '#/definitions/dto.SubjectResponse'
    type: object
...
  dto.RoomResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.SlotResponse:
    properties:
      id:
        type: string
      position:
        type: integer
      queue:
        $ref: '#/definitions/dto.QueueMinimalInfo'
      queue_id:
        type: string
      signed_up_at:
        type: string
      status:
        type: string
      student:
        $ref: '#/definitions/dto.UserMinimalResponse'
    type: object
  dto.SubjectResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
  dto.TeacherResponse:
    properties:
      full_name:
        type: string
      id:
        type: string
    type: object
  dto.TransferHeadmanRoleRequest:
    properties:
      to_user_id:
        type: string
    required:
    - to_user_id
    type: object
  dto.TransferHeadmanRoleResponse:
    properties:
      group_id:
        type: string
      new_headman_id:
        type: string
      previous_headman_id:
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      first_name:
        type: string
      group_id:
        type: string
      last_name:
        type: string
    type: object
...
  dto.UserMinimalResponse:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      group:
        $ref: '#/definitions/dto.GroupResponse'
      id:
        type: string
      is_blocked:
        type: boolean
      last_name:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: HTTP API for Stud Hub backend.
  title: Stud Hub API
  version: "1.0"
paths:
...
  /lessons:
    get:
      description: Returns lessons by group and date range.
      parameters:
      - description: Group ID (UUID). If omitted, current user's group is used.
        in: query
        name: group_id
        type: string
      - description: Start date in YYYY-MM-DD format
        in: query
        name: date_from
        required: true
        type: string
      - description: End date in YYYY-MM-DD format
        in: query
        name: date_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.LessonResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: List lessons
      tags:
      - Lessons
  /lessons/{id}:
    get:
      description: Returns lesson details by lesson ID.
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LessonResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get lesson by ID
      tags:
      - Lessons
  /lessons/imports:
    post:
      consumes:
      - multipart/form-data
      description: Imports lessons from uploaded JSON file (admin endpoint when enabled
        in routing).
      parameters:
      - description: JSON file with import payload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ImportScheduleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Import schedule
      tags:
      - Lessons
...
  /users/me:
    get:
      description: Returns profile for authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get current user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates profile fields for authenticated user.
      parameters:
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update current user
      tags:
      - Users
  /users/me/headman-role:
    put:
      consumes:
      - application/json
      description: Transfers headman role to another user in the same group.
      parameters:
      - description: Transfer headman role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransferHeadmanRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TransferHeadmanRoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Transfer headman role
      tags:
      - Users
  /users/me/slots:
    get:
      description: Returns authenticated user's slots with optional filters.
      parameters:
      - description: Slot status filter
        in: query
        name: status
        type: string
      - description: Queue status filter
        in: query
        name: queue_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SlotResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get current user slots
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"